backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
python-dotenv==1.0.1
earthengine-api==1.1.5
google-generativeai==0.8.3
numpy==1.26.4
scikit-learn==1.5.2
Pillow==11.0.0
matplotlib==3.9.2
